Refactor ThreadPool creation
