+++ src/Microsoft.Diagnostics.Runtime/ClrThreadPool.cs
@@ -18,7 +18,7 @@ namespace Microsoft.Diagnostics.Runtime
     public sealed class ClrThreadPool : IClrThreadPool
     {
         private readonly ClrRuntime _runtime;
-        private readonly IClrThreadPoolHelper _helpers;
+        private readonly IClrThreadPoolHelpers? _legacyData;
         private readonly ulong _nativeLogAddress;
         private readonly uint _nativeLogStart;
         private readonly uint _nativeLogSize;
@@ -73,12 +73,14 @@ namespace Microsoft.Diagnostics.Runtime
         private readonly ClrDataAddress _firstLegacyWorkRequest;
         private readonly ClrDataAddress _asyncTimerFunction;
 
-        internal ClrThreadPool(ClrRuntime runtime, IClrThreadPoolHelper helpers)
+        internal ClrThreadPool(ClrRuntime runtime, IClrThreadPoolHelpers? helpers)
         {
             _runtime = runtime;
-            _helpers = helpers;
+            _legacyData = helpers;
 
-            bool hasLegacyData = _helpers.GetLegacyThreadPoolData(out ThreadPoolData tpData, out bool mustBePortable);
+            ThreadPoolData tpData = default;
+            bool mustBePortable = true;
+            bool hasLegacyData = _legacyData is not null && _legacyData.GetLegacyThreadPoolData(out tpData, out mustBePortable);
 
             ClrObject threadPool = GetPortableThreadPool(mustBePortable);
             if (!threadPool.IsNull && threadPool.IsValid)
@@ -135,8 +137,11 @@ namespace Microsoft.Diagnostics.Runtime
         /// does not have them.</returns>
         public IEnumerable<LegacyThreadPoolWorkRequest> EnumerateLegacyWorkRequests()
         {
+            if (_legacyData is null)
+                yield break;
+
             ulong curr = _firstLegacyWorkRequest;
-            while (curr != 0 && _helpers.GetLegacyWorkRequestData(curr, out WorkRequestData workRequestData))
+            while (curr != 0 && _legacyData.GetLegacyWorkRequestData(curr, out WorkRequestData workRequestData))
             {
                 yield return new LegacyThreadPoolWorkRequest()
                 {
